pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mendix-doc-mcp"
version = "0.1.0"
description = "MCP server for searching and browsing Mendix documentation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.27",
    "python-frontmatter>=1.1",
    "pydantic>=2.0",
    "click>=8.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]

[project.scripts]
mendix-doc-mcp = "src.server.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
